Move FulllyBoundedTimestampRange
It will be needed in atlasdb-cassandra
