+++ atlasdb-backup/src/test/java/com/palantir/atlasdb/backup/transaction/Transactions1TableInteractionTest.java
@@ -32,6 +32,7 @@ import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraConstants;
 import com.palantir.atlasdb.transaction.encoding.V1EncodingStrategy;
 import com.palantir.atlasdb.transaction.impl.TransactionConstants;
+import com.palantir.timestamp.FullyBoundedTimestampRange;
 import java.nio.ByteBuffer;
 import org.apache.commons.codec.binary.Hex;
 import org.junit.Test;
