+++ timestamp-impl/build.gradle
@@ -19,6 +19,8 @@ dependencies {
 
   annotationProcessor 'com.google.auto.service:auto-service'
   compileOnly 'com.google.auto.service:auto-service'
+  annotationProcessor 'org.immutables:value'
+  compileOnly 'org.immutables:value:annotations'
   annotationProcessor project(":atlasdb-processors")
   compileOnly project(":atlasdb-processors")
